+++ example/v9.x.x/dat/src/i18n.js
@@ -1,6 +1,6 @@
 import i18n from 'i18next';
 import LanguageDetector from 'i18next-browser-languagedetector';
-import XHR from 'i18next-xhr-backend';
+import Backend from 'i18next-http-backend';
 import LocizeBackend from 'i18next-locize-backend';
 import whichBackend from './whichBackend';
 
@@ -38,7 +38,7 @@ switch (whichBackend()) {
     options.backend = {
       loadPath: '/locales/{{lng}}/{{ns}}.json',
     };
-    i18n.use(XHR);
+    i18n.use(Backend);
     break;
 
   case 'memory':