+++ src/Trans.js
@@ -1,10 +1,13 @@
 import React, { useContext } from 'react';
 import HTML from 'html-parse-stringify2';
-import { getI18n, getHasUsedI18nextProvider, I18nContext, getDefaults } from './context';
+import { getI18n, I18nContext, getDefaults } from './context';
 import { warn, warnOnce } from './utils';
 
-function hasChildren(node) {
-  return node && (node.children || (node.props && node.props.children));
+function hasChildren(node, checkLength) {
+  if (!node) return false;
+  const base = node.props ? node.props.children : node.children;
+  if (checkLength) return base.length > 0;
+  return !!base;
 }
 
 function getChildren(node) {
@@ -21,55 +24,62 @@ function getAsArray(data) {
   return Array.isArray(data) ? data : [data];
 }
 
-export function nodesToString(startingString, children, index, i18nOptions) {
-  if (!children) return '';
-  let stringNode = startingString;
+function mergeProps(source, target) {
+  const newTarget = { ...target };
+  // overwrite source.props when target.props already set
+  newTarget.props = Object.assign(source.props, target.props);
+  return newTarget;
+}
 
+export function nodesToString(children, i18nOptions) {
+  if (!children) return '';
+  let stringNode = '';
+
+  // do not use `React.Children.toArray`, will fail at object children
   const childrenArray = getAsArray(children);
   const keepArray = i18nOptions.transKeepBasicHtmlNodesFor || [];
 
-  childrenArray.forEach((child, i) => {
-    const elementKey = `${i}`;
-
+  // e.g. lorem <br/> ipsum {{ messageCount, format }} dolor <strong>bold</strong> amet
+  childrenArray.forEach((child, childIndex) => {
     if (typeof child === 'string') {
-      stringNode = `${stringNode}${child}`;
-    } else if (hasChildren(child)) {
-      const elementTag =
-        keepArray.indexOf(child.type) > -1 &&
-        Object.keys(child.props).length === 1 &&
-        typeof hasChildren(child) === 'string'
-          ? child.type
-          : elementKey;
+      // actual e.g. lorem
+      // expected e.g. lorem
+      stringNode += `${child}`;
+    } else if (React.isValidElement(child)) {
+      const childPropsCount = Object.keys(child.props).length;
+      const shouldKeepChild = keepArray.indexOf(child.type) > -1;
+      const childChildren = child.props.children;
 
-      if (child.props && child.props.i18nIsDynamicList) {
-        // we got a dynamic list like "<ul>{['a', 'b'].map(item => ( <li key={item}>{item}</li> ))}</ul>""
-        // the result should be "<0></0>" and not "<0><0>a</0><1>b</1></0>"
-        stringNode = `${stringNode}<${elementTag}></${elementTag}>`;
+      if (!childChildren && shouldKeepChild && childPropsCount === 0) {
+        // actual e.g. lorem <br/> ipsum
+        // expected e.g. lorem <br/> ipsum
+        stringNode += `<${child.type}/>`;
+      } else if (!childChildren && (!shouldKeepChild || childPropsCount !== 0)) {
+        // actual e.g. lorem <hr className="test" /> ipsum
+        // expected e.g. lorem <0></0> ipsum
+        stringNode += `<${childIndex}></${childIndex}>`;
+      } else if (child.props.i18nIsDynamicList) {
+        // we got a dynamic list like
+        // e.g. <ul i18nIsDynamicList>{['a', 'b'].map(item => ( <li key={item}>{item}</li> ))}</ul>
+        // expected e.g. "<0></0>", not e.g. "<0><0>a</0><1>b</1></0>"
+        stringNode += `<${childIndex}></${childIndex}>`;
+      } else if (shouldKeepChild && childPropsCount === 1 && typeof childChildren === 'string') {
+        // actual e.g. dolor <strong>bold</strong> amet
+        // expected e.g. dolor <strong>bold</strong> amet
+        stringNode += `<${child.type}>${childChildren}</${child.type}>`;
       } else {
         // regular case mapping the inner children
-        stringNode = `${stringNode}<${elementTag}>${nodesToString(
-          '',
-          getChildren(child),
-          i + 1,
-          i18nOptions,
-        )}</${elementTag}>`;
-      }
-    } else if (React.isValidElement(child)) {
-      if (keepArray.indexOf(child.type) > -1 && Object.keys(child.props).length === 0) {
-        stringNode = `${stringNode}<${child.type}/>`;
-      } else {
-        stringNode = `${stringNode}<${elementKey}></${elementKey}>`;
+        const content = nodesToString(childChildren, i18nOptions);
+        stringNode += `<${childIndex}>${content}</${childIndex}>`;
       }
     } else if (typeof child === 'object') {
-      const clone = { ...child };
-      const { format } = clone;
-      delete clone.format;
-
+      // e.g. lorem {{ value, format }} ipsum
+      const { format, ...clone } = child;
       const keys = Object.keys(clone);
-      if (format && keys.length === 1) {
-        stringNode = `${stringNode}{{${keys[0]}, ${format}}}`;
-      } else if (keys.length === 1) {
-        stringNode = `${stringNode}{{${keys[0]}}}`;
+
+      if (keys.length === 1) {
+        const value = format ? `${keys[0]}, ${format}` : keys[0];
+        stringNode += `{{${value}}}`;
       } else {
         // not a valid interpolation object (can only contain one value plus format)
         warn(
@@ -125,50 +135,92 @@ function renderNodes(children, targetString, i18n, i18nOptions, combinedTOpts) {
   // -> avoids issues in parser removing prepending text nodes
   const ast = HTML.parse(`<0>${interpolatedString}</0>`);
 
-  function mapAST(reactNode, astNode) {
+  function renderInner(child, node, rootReactNode) {
+    const childs = getChildren(child);
+    const mappedChildren = mapAST(childs, node.children, rootReactNode);
+    // console.warn('INNER', node.name, node, child, childs, node.children, mappedChildren);
+    return hasValidReactChildren(childs) && mappedChildren.length === 0 ? childs : mappedChildren;
+  }
+
+  function pushTranslatedJSX(child, inner, mem, i) {
+    if (child.dummy) child.children = inner; // needed on preact!
+    mem.push(React.cloneElement(child, { ...child.props, key: i }, inner));
+  }
+
+  // reactNode (the jsx root element or child)
+  // astNode (the translation string as html ast)
+  // rootReactNode (the most outer jsx children array or trans components prop)
+  function mapAST(reactNode, astNode, rootReactNode) {
     const reactNodes = getAsArray(reactNode);
     const astNodes = getAsArray(astNode);
 
     return astNodes.reduce((mem, node, i) => {
       const translationContent = node.children && node.children[0] && node.children[0].content;
       if (node.type === 'tag') {
-        const child = reactNodes[parseInt(node.name, 10)] || {};
+        let tmp = reactNodes[parseInt(node.name, 10)]; // regular array (components or children)
+        if (!tmp && rootReactNode.length === 1 && rootReactNode[0][node.name])
+          tmp = rootReactNode[0][node.name]; // trans components is an object
+        if (!tmp) tmp = {};
+        //  console.warn('TMP', node.name, parseInt(node.name, 10), tmp, reactNodes);
+        const child =
+          Object.keys(node.attrs).length !== 0 ? mergeProps({ props: node.attrs }, tmp) : tmp;
+
         const isElement = React.isValidElement(child);
+
+        const isValidTranslationWithChildren =
+          isElement && hasChildren(node, true) && !node.voidElement;
+
+        const isEmptyTransWithHTML =
+          emptyChildrenButNeedsHandling && typeof child === 'object' && child.dummy && !isElement;
+
+        const isKnownComponent =
+          typeof children === 'object' &&
+          children !== null &&
+          Object.hasOwnProperty.call(children, node.name);
+        // console.warn('CHILD', node.name, node, isElement, child);
 
         if (typeof child === 'string') {
           mem.push(child);
-        } else if (hasChildren(child)) {
-          const childs = getChildren(child);
-          const mappedChildren = mapAST(childs, node.children);
-          const inner =
-            hasValidReactChildren(childs) && mappedChildren.length === 0 ? childs : mappedChildren;
-
-          if (child.dummy) child.children = inner; // needed on preact!
-          mem.push(React.cloneElement(child, { ...child.props, key: i }, inner));
         } else if (
-          emptyChildrenButNeedsHandling &&
-          typeof child === 'object' &&
-          child.dummy &&
-          !isElement
+          hasChildren(child) || // the jsx element has children -> loop
+          isValidTranslationWithChildren // valid jsx element with no children but the translation has -> loop
         ) {
+          const inner = renderInner(child, node, rootReactNode);
+          pushTranslatedJSX(child, inner, mem, i);
+        } else if (isEmptyTransWithHTML) {
           // we have a empty Trans node (the dummy element) with a targetstring that contains html tags needing
           // conversion to react nodes
           // so we just need to map the inner stuff
-          const inner = mapAST(reactNodes /* wrong but we need something */, node.children);
+          const inner = mapAST(
+            reactNodes /* wrong but we need something */,
+            node.children,
+            rootReactNode,
+          );
           mem.push(React.cloneElement(child, { ...child.props, key: i }, inner));
         } else if (Number.isNaN(parseFloat(node.name))) {
-          if (i18nOptions.transSupportBasicHtmlNodes && keepArray.indexOf(node.name) > -1) {
+          if (isKnownComponent) {
+            const inner = renderInner(child, node, rootReactNode);
+            pushTranslatedJSX(child, inner, mem, i);
+          } else if (i18nOptions.transSupportBasicHtmlNodes && keepArray.indexOf(node.name) > -1) {
             if (node.voidElement) {
               mem.push(React.createElement(node.name, { key: `${node.name}-${i}` }));
             } else {
-              const inner = mapAST(reactNodes /* wrong but we need something */, node.children);
+              const inner = mapAST(
+                reactNodes /* wrong but we need something */,
+                node.children,
+                rootReactNode,
+              );
 
               mem.push(React.createElement(node.name, { key: `${node.name}-${i}` }, inner));
             }
           } else if (node.voidElement) {
             mem.push(`<${node.name} />`);
           } else {
-            const inner = mapAST(reactNodes /* wrong but we need something */, node.children);
+            const inner = mapAST(
+              reactNodes /* wrong but we need something */,
+              node.children,
+              rootReactNode,
+            );
 
             mem.push(`<${node.name}>${inner}</${node.name}>`);
           }
@@ -197,7 +249,7 @@ function renderNodes(children, targetString, i18n, i18nOptions, combinedTOpts) {
   // call mapAST with having react nodes nested into additional node like
   // we did for the string ast from translation
   // return the children of that extra node to get expected result
-  const result = mapAST([{ dummy: true, children }], ast);
+  const result = mapAST([{ dummy: true, children }], ast, getAsArray(children || []));
   return getChildren(result[0]);
 }
 
@@ -215,11 +267,9 @@ export function Trans({
   t: tFromProps,
   ...additionalProps
 }) {
-  const ReactI18nContext = useContext(I18nContext);
-  const { i18n: i18nFromContext, defaultNS: defaultNSFromContext } = getHasUsedI18nextProvider()
-    ? ReactI18nContext || {}
-    : {};
+  const { i18n: i18nFromContext, defaultNS: defaultNSFromContext } = useContext(I18nContext) || {};
   const i18n = i18nFromProps || i18nFromContext || getI18n();
+
   if (!i18n) {
     warnOnce('You will need pass in an i18next instance by using i18nextReactModule');
     return children;
@@ -228,7 +278,6 @@ export function Trans({
   const t = tFromProps || i18n.t.bind(i18n) || (k => k);
 
   const reactI18nextOptions = { ...getDefaults(), ...(i18n.options && i18n.options.react) };
-  const useAsParent = parent !== undefined ? parent : reactI18nextOptions.defaultTransParent;
 
   // prepare having a namespace
   let namespaces = ns || t.ns || defaultNSFromContext || (i18n.options && i18n.options.defaultNS);
@@ -236,8 +285,9 @@ export function Trans({
 
   const defaultValue =
     defaults ||
-    nodesToString('', children, 0, reactI18nextOptions) ||
-    reactI18nextOptions.transEmptyNodeValue;
+    nodesToString(children, reactI18nextOptions) ||
+    reactI18nextOptions.transEmptyNodeValue ||
+    i18nKey;
   const { hashTransKey } = reactI18nextOptions;
   const key = i18nKey || (hashTransKey ? hashTransKey(defaultValue) : defaultValue);
   const interpolationOverride = values ? {} : { interpolation: { prefix: '#$?', suffix: '?$#' } };
@@ -251,18 +301,17 @@ export function Trans({
   };
   const translation = key ? t(key, combinedTOpts) : defaultValue;
 
-  if (!useAsParent)
-    return renderNodes(
-      components || children,
-      translation,
-      i18n,
-      reactI18nextOptions,
-      combinedTOpts,
-    );
-
-  return React.createElement(
-    useAsParent,
-    additionalProps,
-    renderNodes(components || children, translation, i18n, reactI18nextOptions, combinedTOpts),
+  const content = renderNodes(
+    components || children,
+    translation,
+    i18n,
+    reactI18nextOptions,
+    combinedTOpts,
   );
+
+  // allows user to pass `null` to `parent`
+  // and override `defaultTransParent` if is present
+  const useAsParent = parent !== undefined ? parent : reactI18nextOptions.defaultTransParent;
+
+  return useAsParent ? React.createElement(useAsParent, additionalProps, content) : content;
 }